backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python-headless
scipy
soundfile
pydantic
httpx
PyJWT[crypto]
cryptography
# ML stack — DINOv2 (image), Whisper (audio), MTCNN (face cropping).
# CPU-only torch; install matching CUDA wheels manually if a GPU is available.
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
transformers
facenet-pytorch
# Model weight management
huggingface_hub
# Optional: provenance reading. Comment out if the install fails on your platform.
c2pa-python
